+++ js/src/forum/index.ts
@@ -1,0 +1,41 @@
+import TwoFactorLogInModal from "./components/TwoFactorLogInModal";
+import TwoFactorSetupModal from "./components/TwoFactorSetupModal";
+import TwoFactorState from "./states/TwoFactorState";
+import Button from "flarum/common/components/Button";
+import { extend, override } from "flarum/common/extend";
+import app from "flarum/forum/app";
+import LogInModal from "flarum/forum/components/LogInModal";
+import SettingsPage from "flarum/forum/components/SettingsPage";
+
+app.initializers.add("nearata-twofactor", () => {
+    extend(SettingsPage.prototype, "accountItems", function (items) {
+        items.add(
+            "nearataTwoFactor",
+            m(
+                Button,
+                {
+                    class: "Button",
+                    onclick: () =>
+                        app.modal.show(TwoFactorSetupModal, {
+                            twoFactorState: new TwoFactorState(),
+                        }),
+                },
+                app.translator.trans("nearata-twofactor.forum.setup_button")
+            )
+        );
+    });
+
+    override(LogInModal.prototype, "onerror", function (original, error) {
+        const response = error.response;
+
+        if (response && "has2FA" in response && response.has2FA) {
+            app.modal.show(TwoFactorLogInModal, {
+                identification: this.identification(),
+                password: this.password(),
+                remember: this.remember(),
+            });
+        } else {
+            return original(error);
+        }
+    });
+});