+++ src/Forum/Controller/CustomLogInController.php
@@ -2,7 +2,6 @@
 
 namespace Nearata\TwoFactor\Forum\Controller;
 
-use Flarum\Api\Controller\CreateTokenController;
 use Flarum\Forum\Controller\LogInController;
 use Flarum\Http\AccessToken;
 use Flarum\Http\RememberAccessToken;
@@ -37,14 +36,15 @@ class CustomLogInController extends LogInController
                 }
 
                 $otp = TOTP::create($user->twofa_secret);
-                if (!$otp->verify($code) && !Helpers::isBackupCode($user, $code)) {
+
+                if (!($otp->verify($code) || Helpers::isBackupCode($user, $code))) {
                     return new EmptyResponse(401);
                 }
             }
 
             $token = AccessToken::findValid($data->token);
-
             $session = $request->getAttribute('session');
+
             $this->authenticator->logIn($session, $token);
 
             $this->events->dispatch(new LoggedIn($this->users->findOrFail($data->userId), $token));