+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add php 8 support ([Issue #7](https://github.com/Nearata/flarum-ext-twofactor/issues/7)).
+
 ## [2.0.1] - 2021-09-11
 
-- Fixes [#3](https://github.com/Nearata/flarum-ext-twofactor/issues/3)
+- Fix wrong generation of backup qr codes ([Issue #3](https://github.com/Nearata/flarum-ext-twofactor/issues/3))
 
 ## [2.0.0] - 2021-06-20